Remove timeout and handle store and scroll async
There was a timeout necessary, because otherwise the list could scroll before the new expense is added. Using a timeout was not the best solution here, because we cannot be sure, that a new expense is added always within 200ms. It is better to wait for a new expense to be added. This is achieved now with async-await.